refactor: rename api.device.$deviceId route
