+++ app/routes/api.boxes.$deviceId.ts
